XASY: Change QRect to QRectF.

+++ GUI/xasyBezierInterface.py
@@ -74,7 +74,7 @@ class InteractiveBezierEditor(InplaceAddObj.InplaceObjProcess):
                 continue
 
             selEpsilon = 6/self.info['magnification']
-            newRect = QtCore.QRect(0, 0, 2 * selEpsilon, 2 * selEpsilon)
+            newRect = QtCore.QRect(0, 0, int(2 * selEpsilon), int(2 * selEpsilon))
             x, y = self.transf * node
             x = int(round(x))
             y = int(round(y))
